+++ src/main.py
@@ -42,6 +42,34 @@ class MDProperty:
         return f"{self.name}"
 
 
+class PYFile:
+    def __init__(self):
+        self.classes = []
+
+
+class PYClass:
+    def __init__(self, name: str):
+        self.name: str = name
+        self.properties: list[PYProperty] = []
+        self.methods: list[PYMethod] = []
+
+    def __str__(self):
+        return f"{self.name}"
+
+
+class PYMethod:
+    def __init__(self, name: str, params):
+        self.name: str = name
+        self.params = params
+
+
+class PYProperty:
+    def __init__(self, name: str, type_: str, value: str):
+        self.name: str = name
+        self.type_: str = type_
+        self.value = value
+
+
 class AttrType(Enum):
     properties = 1;
     methods = 2;
@@ -50,13 +78,15 @@ class AttrType(Enum):
 class Parsnip:
     def __init__(self):
         self.md_file = MDFile()
+        self.py_file = PYFile()
 
         self.attr_store_type = AttrType.properties;
 
         md_data = self.parse_md_file("data/test.md")
         self.store_md_data(md_data)
         
-        self.parse_py_file()
+        py_data = self.parse_py_file("data/test.py")
+        self.store_py_data(py_data)
 
     def parse_md_file(self, filepath: str) -> list[str]:
         with open(filepath, "rb") as md:
@@ -67,12 +97,13 @@ class Parsnip:
 
             return md_data
 
-    def parse_py_file(self):
-        with open("data/test.py", "rb") as py:
+    def parse_py_file(self, filepath: str) -> list[str]:
+        with open(filepath, "rb") as py:
             data = py.read()
             parsed_data = subprocess.check_output(["cabal", "exec", "parsnip", data, "py"], text=True)
+            py_data = json.loads(parsed_data)
 
-            print(parsed_data)
+            return py_data
 
     def store_md_data(self, data: list[str]) -> None:
         for line in data:
@@ -100,7 +131,28 @@ class Parsnip:
                 elif self.attr_store_type == AttrType.methods:
                     self.md_file.classes[-1].methods.append(MDMethod(*rest))
 
+    def store_py_data(self, data: list[str]):
+        for line in data:
+            items = line.split("<s>")
+            item_type = items[0]
+            rest = items[1:]
+
+            if item_type == "Class":
+                self.py_file.classes.append(PYClass(*rest))
+
+            if item_type == "Function":
+                name = rest[0]
+                params = rest[2:-1]
+                self.py_file.classes[-1].methods.append(PYMethod(name, params))
+
+            if item_type == "Property":
+                self.py_file.classes[-1].properties.append(PYProperty(*rest))
+
+
 
 if __name__ == "__main__":
     app = Parsnip()
+
+    for c in app.py_file.classes:
+        print(c)